Mount rootfs from SD card file
We mount the rootfs from a custom initramfs built
into the kernel.

We link BusyBox statically. We install a single
executable to initramfs. /init creates symlinks
to /bin/busybox for each applet.

When using initramfs the kernel does not mount
devtmps automatically. /init mounts it instead.

/sbin/init will not remount the rootfs according
to /etc/fstab. The rootfs will remain as /init
left it.